Export new error APIs from yarpc package
This merges the yarpcerrors package into the api/transport package and
re-exports relevant functions from the top-level YARPC package for use
by service authors. Based on offline discussion, we feel that this
provides a cleaner API than an errors subpackage.
